requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clients==1.4
decorator==4.4.2
Flask==2.2.3
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
icrawler==0.6.6
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
pillar-youtube-upload==0.3.0
Pillow==9.4.0
proglog==0.1.10
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pyparsing==3.0.9
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
router==0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tgclient==1.5.3
tqdm==4.64.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
